Fix module cloning.
- Insert return instruction in the beginning.

Without a terminator, LLVM crashes when trying to clone the module. Therefore, it makes
more sense to create the return instruction at first, and prepend the other instructions.

- Verify before crash

Apparently, `llvm::orc::cloneToNewContext` raises a segmentation fault if we try to clone
an incorrect module. Therefore, we need to verify its (type) correctness first.